Trivial Refactors.
Unnecessary `unsigned' are replaced by simply longer type.  And branches in the
state macine are more clear now.

+++ config.h
@@ -11,4 +11,4 @@ mod_key mod_map[] =
 
 // If a key is held down for a time greater than max_delay, then,
 // when released, it will not send its primary function.
-unsigned int max_delay = 300;
+long max_delay = 300;
